requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.7.0
aiohttp>=3.9.0
lxml>=4.9.0
tmdbv3api>=1.7.0
transformers>=4.40.0
torch>=2.3.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0